refactor(settings): split training restore page into its own settings page
Move the training restore UI and logic out of the main settings page into a dedicated /settings/restoreTraining page. Added renderer/pages/settings/restoreTraining.tsx which contains the restore modal, localStorage lookup (training-backup) and navigation to /trainings/active. Updated renderer/pages/settings/index.tsx to remove restore-related UI, imports and hooks and to focus on database settings. Updated renderer/lib/constants.ts to add a new SETTINGS_ROUTES entry for the restore page and adjust the main settings label accordingly. This separates concerns and cleans up unused imports/types in the main settings page.
